Added new *.java to Makefile.in, added a utility class.
