+++ src/graph/Plot.java
@@ -28,6 +28,7 @@ import net.opentsdb.core.Const;
 import net.opentsdb.core.DataPoint;
 import net.opentsdb.core.DataPoints;
 import net.opentsdb.meta.Annotation;
+import net.opentsdb.utils.FileSystem;
 
 /**
  * Produces files to generate graphs with Gnuplot.
@@ -196,7 +197,7 @@ public final class Plot {
     int npoints = 0;
     final int nseries = datapoints.size();
     final String datafiles[] = nseries > 0 ? new String[nseries] : null;
-    checkDirectory(new File(basepath).getParent(),
+    FileSystem.checkDirectory(new File(basepath).getParent(),
         Const.MUST_BE_WRITEABLE, Const.CREATE_IF_NEEDED);
     for (int i = 0; i < nseries; i++) {
       datafiles[i] = basepath + "_" + i + ".dat";
@@ -395,32 +396,6 @@ public final class Plot {
       return "%b %d";
     } else {
       return "%Y/%m/%d";
-    }
-  }
-
-  /**
-   * Verifies a directory and checks to see if it's writeable or not if
-   * configured
-   * @param dir The path to check on
-   * @param need_write Set to true if the path needs write access
-   * @param create Set to true if the directory should be created if it does not
-   *          exist
-   * @throws IllegalArgumentException if the path is empty, if it's not there
-   *           and told not to create it or if it needs write access and can't
-   *           be written to
-   */
-  public static void checkDirectory(final String dir,
-      final boolean need_write, final boolean create) {
-    if (dir.isEmpty())
-      throw new IllegalArgumentException("Directory path is empty");
-    final File f = new File(dir);
-    if (!f.exists() && !(create && f.mkdirs())) {
-      throw new IllegalArgumentException("No such directory [" + dir + "]");
-    } else if (!f.isDirectory()) {
-      throw new IllegalArgumentException("Not a directory [" + dir + "]");
-    } else if (need_write && !f.canWrite()) {
-      throw new IllegalArgumentException("Cannot write to directory [" + dir
-          + "]");
     }
   }
 }
